requirements.txt
flask
flask_cors
boto3
pymongo
pg8000
psycopg2
nats-py
nest_asyncio

